app/requirements.txt
uvicorn
fastapi
pydantic[email]
httpx

singleton_package
json-advanced
python-dotenv
debugpy
ipython

aiofiles
aiocache

beanie
fastapi-mongo-base
ufaas-fastapi-business

apscheduler
pillow
pytz

usso[fastapi]
ufaas
ufiles

selenium
beautifulsoup4
google-api-python-client
